correctifs

+++ 01-initial-install.sh
@@ -49,25 +49,19 @@ sed -i -e 's/'"$ActualFullHostname"'/'"$fqdn"'/' '/etc/hosts'
 sed -i -e 's/'"$ActualServerName"'/'"$server_name"'/' '/etc/hosts'
 
 
-apt -y install sudo                          #
-sudo apt update -y
-sudo apt upgrade -y
+read -p "Voulez-vous lancer la mise-a-jour du systeme [O/n] ? " update_choice
+if [[ "$update_choice" = 'O' ]]; then
+  apt -y install sudo                          #
+  sudo apt update -y
+  sudo apt upgrade -y
+fi
 sleep 8
 
 echo ""
-echo -e "${GREEN}### Installation of zippers and unzippers${NOCOLOR}"
+echo -e "${GREEN}### Installation des paquets necessaires${NOCOLOR}"
 
-sudo apt -y install unrar-free unzip zip                             # Archiver for .rar files
+sudo apt -y install unrar-free unzip zip hardinfo hwinfo htop sysv-rc-conf locate git curl iperf                 
 sleep 8
-
-echo ""
-echo -e "${GREEN}### System utilities and various tools${NOCOLOR}"
-
-sudo apt -y install hardinfo hwinfo htop sysv-rc-conf locate git         # Displays system information
-
-# Core tools
-
-sudo apt -y install curl iperf                           # Command line tool for transferring data with URL syntax
 
 echo ""
 echo -e "${GREEN}### Création de l'utilisateur${NOCOLOR}"
@@ -90,53 +84,65 @@ git config --global color.branch auto
 read -p "Utilisateur Git : " git_user
 read -p "Email pour Git : " git_email
 git config --global user.name $git_user
-git confit --global user.mail $git_email
+git config --global user.mail $git_email
 cd /opt/
-  if [ ! -d /opt/Git_Repos/cybit_script ]; then
+  if [ ! -d /opt/Git_Repos/cybit-script ]; then
     mkdir -p /opt/Git_Repos/
     git clone https://github.com/petitsurfeur/cybit-script.git
   else echo -e "Le dossier /opt/Git_Repos/cybit_script/ existe deja"
   fi
 
-cd /opt/Git_Repos/cybit_script/update/ && install_update.sh
+cd /opt/Git_Repos/cybit-script/update/ && ./install_update.sh
+
+echo -e "${GREEN}### Installation et configuration des outils tiers${NOCOLOR}"
 
 read -p "Voulez-vous installer Vim [O-n) ? " vim_choice
   if [[ "$vim_choice" = 'O' ]]; then
-        cd /opt/Git_Repos/cybit_script/vim/ && ./install_vim.sh
+        cd /opt/Git_Repos/cybit-script/vim/ && ./install_vim.sh
           fi
 
 read -p "Voulez-vous installer SSH [O-n) ? " ssh_choice
   if [[ "$ssh_choice" = 'O' ]]; then
-        cd /opt/Git_Repos/cybit_script/ssh/ && ./install_ssh.sh
+        cd /opt/Git_Repos/cybit-script/ssh/ && ./install_ssh.sh
 	  fi
 
 read -p "Voulez-vous installer .bashrc [O-n) ? " bashrc_choice
   if [[ "$bashrc_choice" = 'O' ]]; then
-        cd /opt/Git_Repos/cybit_script/bashrc && ./install_bashrc.sh
+        cd /opt/Git_Repos/cybit-script/bashrc && ./install_bashrc.sh
 	  fi
 
 read -p "Voulez-vous installer Exim4 [O-n) ? " exim4_choice
   if [[ "$exim4_choice" = 'O' ]]; then
-        cd /opt/Git_Repos/cybit_script/exim4 && ./install_exim4.sh
+        cd /opt/Git_Repos/cybit-script/exim4 && ./install_exim4.sh
 	  fi
 
 read -p "Voulez-vous installer Apticron [O-n) ? " apticron_choice
   if [[ "$apticron_choice" = 'O' ]]; then
-        cd /opt/Git_Repos/cybit_script/apticron && ./install_apticron.sh
+        cd /opt/Git_Repos/cybit-script/apticron && ./install_apticron.sh
 	  fi
 
 read -p "Voulez-vous installer Fail2ban [O-n) ? " fail2ban_choice
   if [[ "$fail2ban_choice" = 'O' ]]; then
-         cd /opt/Git_Repos/cybit_script/fail2ban && ./install_fail2ban.sh
+         cd /opt/Git_Repos/cybit-script/fail2ban && ./install_fail2ban.sh
            fi
 
 read -p "Voulez-vous installer  UFW [O-n) ? " ufw_choice
   if [[ "$ufw_choice" = 'O' ]]; then
-         cd /opt/Git_Repos/cybit_script/firewall && ./install_firewall.sh
+         cd /opt/Git_Repos/cybit-script/firewall && ./install_firewall.sh
            fi
 
 read -p "Voulez-vous installer OpenVPN [O-n) ? " openvpn_choice
   if [[ "$openvpn_choice" = 'O' ]]; then
-         cd /opt/Git_Repos/cybit_script/openvpn && ./install_openvpn_server.sh
+         cd /opt/Git_Repos/cybit-script/openvpn && ./install_openvpn_server.sh
            fi
+
+echo ""
+echo -e "${GREEN}################################################################"
+echo "###           Installation terminee                          ###"
+echo -e "################################################################${NOCOLOR}"
+echo ""
+read -p "Voulez-vous redemarrer [O/n] ? " reboot
+if [[ "$reboot" = 'O' ]]; then
+  shutdown -r now
+fi
